fin_pos algorithm

+++ math.c
@@ -6,20 +6,20 @@
 /*   By: chon <chon@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/22 13:01:26 by chon              #+#    #+#             */
-/*   Updated: 2024/05/27 14:06:55 by chon             ###   ########.fr       */
+/*   Updated: 2024/05/29 14:57:34 by chon             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "push_swap.h"
 
-int	min(int num1, int num2)
+long	min(long num1, long num2)
 {
 	if (num1 < num2)
 		return (num1);
 	return (num2);
 }
 
-int	max(int num1, int num2)
+long	max(long num1, long num2)
 {
 	if (num1 > num2)
 		return (num1);
